refactor: 알림 관련 미사용 클래스 제거 & 패키지 이동
