+++ backend/src/main/java/mouda/backend/chat/implement/notification/ChatNotificationEvent.java
@@ -1,8 +1,10 @@
-package mouda.backend.chat.domain;
+package mouda.backend.chat.implement.notification;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import mouda.backend.chat.domain.Chat;
+import mouda.backend.chat.domain.ChatRoom;
 
 @Getter
 @RequiredArgsConstructor
